+++ connectors/adapters/activemq/activemq-connector/src/main/java/org/platformlambda/activemq/ArtemisConnector.java
@@ -45,23 +45,68 @@ import javax.jms.JMSException;
 import java.io.IOException;
 import java.util.List;
 import java.util.Properties;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
 
 @CloudConnector(name="activemq")
 public class ArtemisConnector implements CloudSetup {
     private static final Logger log = LoggerFactory.getLogger(ArtemisConnector.class);
 
+    private static final String SYSTEM = "system";
+    private static final String CLOUD_CLIENT_PROPERTIES = "cloud.client.properties";
     public static final String BROKER_URL = "bootstrap.servers";
     private static final String CLOUD_CONNECTOR_HEALTH = "cloud.connector.health";
     private static final String USER_ID = "user.id";
     private static final String USER_PWD = "user.password";
-    private static Properties properties;
-    private static Connection connection;
 
-    public static Properties getClusterProperties() {
+    private static final ConcurrentMap<String, Connection> allConnections = new ConcurrentHashMap<>();
+    private static final ConcurrentMap<String, Properties> allProperties = new ConcurrentHashMap<>();
+
+    public static Properties getClusterProperties(String location) {
+        // default location is cloud.client.properties
+        Properties properties = allProperties.get(location);
+        if (properties == null) {
+            ConfigReader clusterConfig = null;
+            try {
+                clusterConfig = ConnectorConfig.getConfig(location,
+                        "file:/tmp/config/activemq.properties,classpath:/activemq.properties");
+            } catch (IOException e) {
+                log.error("Unable to find activemq properties - {}", e.getMessage());
+                System.exit(-1);
+            }
+            properties = new Properties();
+            for (String k : clusterConfig.getMap().keySet()) {
+                properties.setProperty(k, clusterConfig.getProperty(k));
+            }
+            String url = properties.getProperty(BROKER_URL);
+            Utility util = Utility.getInstance();
+            List<String> cluster = util.split(url, ", ");
+            boolean reachable = false;
+            for (String address : cluster) {
+                int start = address.lastIndexOf('/');
+                int colon = address.lastIndexOf(':');
+                if (colon > 1 && colon > start) {
+                    String host = address.substring(start+1, colon);
+                    int port = util.str2int(address.substring(colon + 1));
+                    if (port > 0) {
+                        // ping the address to confirm it is reachable before making a client connection
+                        if (util.portReady(host, port, 10000)) {
+                            reachable = true;
+                        }
+                    }
+                }
+            }
+            if (!reachable) {
+                log.error("ActiveMQ cluster {} is not reachable", cluster);
+                System.exit(-1);
+            }
+            allProperties.put(location, properties);
+        }
         return properties;
     }
 
-    public static synchronized Connection getConnection() throws JMSException {
+    public static synchronized Connection getConnection(String domain, Properties properties) throws JMSException {
+        Connection connection = allConnections.get(domain);
         if (connection == null) {
             String cluster = properties.getProperty(BROKER_URL, "tcp://127.0.0.1:61616");
             String userId = properties.getProperty(USER_ID, "");
@@ -70,25 +115,29 @@ public class ArtemisConnector implements CloudSetup {
             connection = factory.createConnection(userId, password);
             connection.setExceptionListener((e) -> {
                 String error = e.getMessage();
-                log.error("Tibco cluster exception - {}", error);
+                log.error("Activemq cluster exception - {}", error);
                 if (error != null && (error.contains("terminated") || error.contains("disconnect"))) {
-                    ArtemisConnector.stopConnection();
+                    ArtemisConnector.stopConnection(domain);
                     System.exit(10);
                 }
             });
             connection.start();
             log.info("Connection started - {}", cluster);
-            ConnectorConfig.setServiceName("activemq-artemis");
-            ConnectorConfig.setDisplayUrl(cluster);
+            if (SYSTEM.equals(domain)) {
+                ConnectorConfig.setServiceName("activemq-artemis");
+                ConnectorConfig.setDisplayUrl(cluster);
+            }
+            allConnections.put(domain, connection);
         }
         return connection;
     }
 
-    public static synchronized void stopConnection() {
+    public static synchronized void stopConnection(String domain) {
+        Connection connection = allConnections.get(domain);
         if (connection != null) {
             try {
+                allConnections.remove(domain);
                 connection.stop();
-                connection = null;
                 log.info("Connection stopped");
             } catch (JMSException e) {
                 // ok to ignore
@@ -98,46 +147,11 @@ public class ArtemisConnector implements CloudSetup {
 
     @Override
     public void initialize() {
-        Utility util = Utility.getInstance();
-        ConfigReader clusterConfig = null;
-        try {
-            clusterConfig = ConnectorConfig.getConfig("cloud.client.properties",
-                    "file:/tmp/config/activemq.properties,classpath:/activemq.properties");
-        } catch (IOException e) {
-            log.error("Unable to find activemq.properties - {}", e.getMessage());
-            System.exit(-1);
-        }
-        properties = new Properties();
-        for (String k : clusterConfig.getMap().keySet()) {
-            properties.setProperty(k, clusterConfig.getProperty(k));
-        }
-        String url = properties.getProperty(BROKER_URL);
-        ConnectorConfig.setServiceName("activemq-artemis");
-        ConnectorConfig.setDisplayUrl(url);
-        List<String> cluster = util.split(url, ", ");
-        boolean reachable = false;
-        for (String address : cluster) {
-            int start = address.lastIndexOf('/');
-            int colon = address.lastIndexOf(':');
-            if (colon > 1 && colon > start) {
-                String host = address.substring(start+1, colon);
-                int port = util.str2int(address.substring(colon + 1));
-                if (port > 0) {
-                    // ping the address to confirm it is reachable before making a client connection
-                    if (util.portReady(host, port, 10000)) {
-                        reachable = true;
-                    }
-                }
-            }
-        }
-        if (!reachable) {
-            log.error("ActiveMQ cluster {} is not reachable", cluster);
-            System.exit(-1);
-        }
         try {
             Platform platform = Platform.getInstance();
-            PubSub ps = PubSub.getInstance();
-            ps.enableFeature(new PubSubManager());
+            PubSub ps = PubSub.getInstance(SYSTEM);
+            Properties properties = getClusterProperties(CLOUD_CLIENT_PROPERTIES);
+            ps.enableFeature(new PubSubManager(SYSTEM, properties, ServiceRegistry.CLOUD_MANAGER));
             AppConfigReader config = AppConfigReader.getInstance();
             if (!"true".equals(config.getProperty("service.monitor", "false"))) {
                 // start presence connector