+++ connectors/adapters/tibco/tibco-connector/src/main/java/org/platformlambda/tibco/TibcoConnector.java
@@ -50,11 +50,15 @@ import java.io.IOException;
 import java.util.Hashtable;
 import java.util.List;
 import java.util.Properties;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
 
 @CloudConnector(name="tibco")
 public class TibcoConnector implements CloudSetup {
     private static final Logger log = LoggerFactory.getLogger(TibcoConnector.class);
 
+    private static final String SYSTEM = "system";
+    private static final String CLOUD_CLIENT_PROPERTIES = "cloud.client.properties";
     public static final String BROKER_URL = "bootstrap.servers";
     private static final String CLOUD_CONNECTOR_HEALTH = "cloud.connector.health";
     private static final String USER_ID = "user.id";
@@ -64,11 +68,57 @@ public class TibcoConnector implements CloudSetup {
     private static final String JNDI_USER = "jndi.user";
     private static final String JNDI_PWD = "jndi.password";
     private static final String CONNECTION_FACTORY_NAME = "connection.factory.name";
-    private static Properties properties;
-    private static Connection connection;
-    private static TibjmsAdmin adminClient;
 
-    public static synchronized Connection getConnection() throws Exception {
+    private static final ConcurrentMap<String, Connection> allConnections = new ConcurrentHashMap<>();
+    private static final ConcurrentMap<String, TibjmsAdmin> allAdminClients = new ConcurrentHashMap<>();
+    private static final ConcurrentMap<String, Properties> allProperties = new ConcurrentHashMap<>();
+
+    public static synchronized Properties getTibcoProperties(String location) {
+        // default location is cloud.client.properties
+        Properties properties = allProperties.get(location);
+        if (properties == null) {
+            properties = new Properties();
+            ConfigReader config = null;
+            try {
+                config = ConnectorConfig.getConfig(location,
+                        "file:/tmp/config/tibco.properties,classpath:/tibco.properties");
+            } catch (IOException e) {
+                log.error("Unable to find tibco properties - {}", e.getMessage());
+                System.exit(-1);
+            }
+            for (String k : config.getMap().keySet()) {
+                properties.setProperty(k, config.getProperty(k));
+            }
+            final boolean jndi = "true".equalsIgnoreCase(properties.getProperty(JNDI));
+            String url = properties.getProperty(jndi? PROVIDER_URL : BROKER_URL, "tcp://127.0.0.1:7222");
+            Utility util = Utility.getInstance();
+            List<String> cluster = util.split(url, ", ");
+            boolean reachable = false;
+            for (String address : cluster) {
+                int start = address.lastIndexOf('/');
+                int colon = address.lastIndexOf(':');
+                if (colon > 1 && colon > start) {
+                    String host = address.substring(start+1, colon);
+                    int port = util.str2int(address.substring(colon + 1));
+                    if (port > 0) {
+                        // ping the address to confirm it is reachable before making a client connection
+                        if (util.portReady(host, port, 10000)) {
+                            reachable = true;
+                        }
+                    }
+                }
+            }
+            if (!reachable) {
+                log.error("Tibco EMS cluster {} is not reachable", cluster);
+                System.exit(-1);
+            }
+            allProperties.put(location, properties);
+        }
+        return properties;
+    }
+
+    public static synchronized Connection getConnection(String domain, Properties properties) throws Exception {
+        Connection connection = allConnections.get(domain);
         if (connection == null) {
             final ConnectionFactory factory;
             final boolean jndi = "true".equalsIgnoreCase(properties.getProperty(JNDI));
@@ -95,23 +145,27 @@ public class TibcoConnector implements CloudSetup {
                 String error = e.getMessage();
                 log.error("Tibco cluster exception - {}", error);
                 if (error != null && (error.contains("terminated") || error.contains("disconnect"))) {
-                    TibcoConnector.stopConnection();
+                    TibcoConnector.stopConnection(domain);
                     System.exit(10);
                 }
             });
             connection.start();
             log.info("Connection started - {}", cluster);
-            ConnectorConfig.setServiceName("tibco");
-            ConnectorConfig.setDisplayUrl(cluster);
+            if (SYSTEM.equals(domain)) {
+                ConnectorConfig.setServiceName("tibco");
+                ConnectorConfig.setDisplayUrl(cluster);
+            }
+            allConnections.put(domain, connection);
         }
         return connection;
     }
 
-    public static synchronized void stopConnection() {
+    public static synchronized void stopConnection(String domain) {
+        Connection connection = allConnections.get(domain);
         if (connection != null) {
             try {
+                allConnections.remove(domain);
                 connection.stop();
-                connection = null;
                 log.info("Connection stopped");
             } catch (JMSException e) {
                 // ok to ignore
@@ -119,57 +173,26 @@ public class TibcoConnector implements CloudSetup {
         }
     }
 
-    public static synchronized TibjmsAdmin getAdminClient() throws TibjmsAdminException {
-        if (adminClient == null) {
+    public static synchronized TibjmsAdmin getAdminClient(String domain, Properties properties)
+                throws TibjmsAdminException {
+        TibjmsAdmin client = allAdminClients.get(domain);
+        if (client == null) {
             String cluster = properties.getProperty(BROKER_URL, "tcp://127.0.0.1:7222");
             String userId = properties.getProperty(USER_ID, "");
             String password = properties.getProperty(USER_PWD, "");
-            adminClient = new TibjmsAdmin(cluster, userId, password);
+            client = new TibjmsAdmin(cluster, userId, password);
+            allAdminClients.put(domain, client);
         }
-        return adminClient;
+        return client;
     }
 
     @Override
     public void initialize() {
-        Utility util = Utility.getInstance();
-        ConfigReader clusterConfig = null;
-        try {
-            clusterConfig = ConnectorConfig.getConfig("cloud.client.properties",
-                    "file:/tmp/config/tibco.properties,classpath:/tibco.properties");
-        } catch (IOException e) {
-            log.error("Unable to find tibco.properties - {}", e.getMessage());
-            System.exit(-1);
-        }
-        properties = new Properties();
-        for (String k : clusterConfig.getMap().keySet()) {
-            properties.setProperty(k, clusterConfig.getProperty(k));
-        }
-        final boolean jndi = "true".equalsIgnoreCase(properties.getProperty(JNDI));
-        String url = properties.getProperty(jndi? PROVIDER_URL : BROKER_URL, "tcp://127.0.0.1:7222");
-        List<String> cluster = util.split(url, ", ");
-        boolean reachable = false;
-        for (String address : cluster) {
-            int start = address.lastIndexOf('/');
-            int colon = address.lastIndexOf(':');
-            if (colon > 1 && colon > start) {
-                String host = address.substring(start+1, colon);
-                int port = util.str2int(address.substring(colon + 1));
-                if (port > 0) {
-                    // ping the address to confirm it is reachable before making a client connection
-                    if (util.portReady(host, port, 10000)) {
-                        reachable = true;
-                    }
-                }
-            }
-        }
-        if (!reachable) {
-            log.error("Tibco EMS cluster {} is not reachable", cluster);
-            System.exit(-1);
-        }
         try {
             Platform platform = Platform.getInstance();
-            PubSub ps = PubSub.getInstance();
-            ps.enableFeature(new PubSubManager());
+            PubSub ps = PubSub.getInstance(SYSTEM);
+            Properties properties = getTibcoProperties(CLOUD_CLIENT_PROPERTIES);
+            ps.enableFeature(new PubSubManager(SYSTEM, properties, ServiceRegistry.CLOUD_MANAGER));
             AppConfigReader config = AppConfigReader.getInstance();
             if (!"true".equals(config.getProperty("service.monitor", "false"))) {
                 // start presence connector